wip adds mapcard_preview class
