+++ lib/assets/javascripts/cartodb/new_dashboard/maps/maps_item.js
@@ -2,6 +2,7 @@ var cdb = require('cartodb.js');
 var moment = require('moment');
 var navigateThroughRouter = require('new_common/view_helpers/navigate_through_router');
 var MapviewsGraph = require('new_dashboard/mapviews_graph');
+var MapCardPreview = require('new_dashboard/mapcard_preview');
 var LikesView = require('new_common/views/likes/view');
 var Utils = require('cdb.Utils');
 cdb.admin = require('cdb.admin');
@@ -95,53 +96,12 @@ module.exports = cdb.core.View.extend({
   },
 
   _renderMapThumbnail: function() {
-
-    var host = this.router.currentUserUrl.host();
-
-    var vizjson = host + "/api/v2/viz/" + this.model.get("id") + "/viz.json";
-
-    var width  = 300;
-    var height = 170;
-
-    var $header = this.$el.find(".js-header");
-
-    $header.addClass("is-loading");
-
-    cdb.Image(vizjson).size(width, height).getUrl(function(error, url) {
-
-      var onError = function(error) {
-        $header.removeClass("is-loading");
-        $header.addClass("has-error");
-
-        var r = Math.round(Math.random() * 4); // number to get a random placeholder
-        var $error = $('<div class="MapCard-error error-0' + r + '" />');
-        $header.append($error);
-        $error.fadeIn(250);
-      };
-
-      if (error) {
-        onError(error);
-      } else {
-
-        var img = new Image(); 
-
-        img.onerror = function() {
-          onError(error);
-        }
-
-        img.onload  = function() {
-          var $img = $('<img class="MapCard-preview" src="' + url + '" />');
-          $header.append($img);
-          $header.removeClass("is-loading");
-          $img.fadeIn(250);
-        };
-
-        img.src = url;
-
-      }
-
+    var mapCardPreview = new MapCardPreview({
+      router: this.router,
+      model: this.model,
+      el: this.$('.js-header')
     });
-
+    this.addView(mapCardPreview.render());
   },
 
   _onCardClick: function(e) {
